Changes to be committed: 	modified:   port/boards/labplus_Ledong_pro/bsp_audio_board.c 	modified:   port/boards/labplus_Ledong_pro/mpconfigboard.cmake 	modified:   port/drivers/audio/audio.c 	new file:   port/drivers/audio/es8388.c 	new file:   port/drivers/audio/include/audio_hal.h 	modified:   port/drivers/audio/include/bsp_audio.h 	new file:   port/drivers/audio/include/es8388.h 	new file:   port/drivers/audio/include/esxxx_common.h 	modified:   port/drivers/audio/wav_codec.c 	new file:   port/test/1.wav

+++ port/drivers/audio/include/bsp_audio.h
@@ -73,7 +73,7 @@ esp_err_t bsp_audio_play(const int16_t* data, int length, TickType_t ticks_to_wa
  *    - ESP_OK                  Success
  *    - Others                  Fail
  */
-esp_err_t bsp_get_feed_data(bool is_get_raw_channel, int16_t *buffer, int buffer_len);
+esp_err_t bsp_get_feed_data(bool is_get_raw_channel, uint8_t *buffer, int buffer_len);
 
 int bsp_get_feed_channel(void);
